+++ parser-typechecker/src/Unison/Parser.hs
@@ -1,207 +1,266 @@
-{-# Language DeriveFunctor #-}
-{-# Language DeriveTraversable #-}
-{-# Language DeriveFoldable #-}
-{-# Language BangPatterns #-}
-{-# Language ExplicitForAll #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE ViewPatterns #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Unison.Parser where
 
-import Control.Applicative
-import Control.Monad
-import Data.Char (isSpace)
-import Data.List hiding (takeWhile)
-import Data.Maybe
-import Data.Text (Text)
-import Prelude hiding (takeWhile)
-import qualified Unison.UnisonFile as UnisonFile
-import qualified Data.Char as Char
-import qualified Data.Map  as Map
+import           Control.Applicative
+import           Control.Monad (join)
+import           Data.Bifunctor (bimap)
+import           Data.List.NonEmpty (NonEmpty(..))
+import qualified Data.Map as Map
+import           Data.Maybe
+import qualified Data.Set as Set
+import           Data.Text (Text)
 import qualified Data.Text as Text
-import qualified Prelude
-import qualified Text.Parsec as Parsec
-import           Text.Parsec.Prim (ParsecT)
-import qualified Text.Parsec.Layout as L
-
-import Debug.Trace
-import Text.Parsec (anyChar)
+import           Data.Typeable (Proxy(..))
+import           Text.Megaparsec (runParserT)
+import qualified Text.Megaparsec as P
+import qualified Text.Megaparsec.Char as P
+import           Text.Megaparsec.Error (ShowErrorComponent(..))
+import qualified Unison.ABT as ABT
+import           Unison.Hash
+import qualified Unison.Lexer as L
+import           Unison.Pattern (PatternP)
+import qualified Unison.PatternP as Pattern
+import           Unison.Term (MatchCase(..))
+import qualified Unison.UnisonFile as UnisonFile
+import           Unison.Var (Var)
+import qualified Unison.Var as Var
 
 type PEnv = UnisonFile.CtorLookup
+type P v = P.ParsecT (Error v) Input ((->) PEnv)
+type Token s = P.Token s
+type Err v = P.ParseError (Token Input) (Error v)
+
+data Error v
+  = SignatureNeedsAccompanyingBody (L.Token v)
+  -- we would include the last binding term if we didn't have to have an Ord instance for it
+  | BlockMustEndWithExpression { blockAnn :: Ann, lastBindingAnn :: Ann }
+  | EmptyBlock (L.Token String)
+  | UnknownEffectConstructor (L.Token String)
+  | UnknownDataConstructor (L.Token String)
+  deriving (Show, Eq, Ord)
+
+instance Var v => ShowErrorComponent (Error v) where
+  showErrorComponent e = case e of
+    SignatureNeedsAccompanyingBody t ->
+      showLineCol t ++ ": You provided a type signature, but I " ++
+      "didn't find an accompanying definition after it."
+    BlockMustEndWithExpression bann lbann ->
+      showLineCol lbann ++
+      ": The last line of the block starting at " ++
+      showLineCol bann ++
+      " has to be an expression, not a binding or an import."
+    EmptyBlock t ->
+      showLineCol t ++
+      ": I expected a block after `" ++ L.payload t ++
+      "`, but there wasn't one. Maybe check your indentation."
+    UnknownEffectConstructor t ->
+      showLineCol t ++
+      ": I don't know about any effect constructor named `" ++ L.payload t ++
+      "`. Maybe make sure it's correctly spelled, and that you've imported it."
+    UnknownDataConstructor t ->
+      showLineCol t ++
+      ": I don't know about any data constructor named `" ++ L.payload t ++
+      "`. Maybe make sure it's correctly spelled, and that you've imported it."
+
+data Ann =
+  Intrinsic |
+  Ann { start :: L.Pos, end :: L.Pos } deriving (Eq, Ord, Show)
+
+instance Semigroup Ann where
+  Ann s1 _ <> Ann _ e2 = Ann s1 e2
+  Intrinsic <> Intrinsic = error "FUN SURPRISE!"
+  x <> y = error $ "Compiler bug! Tried to combine terms annotated with ("
+                   ++ show x ++ ") and (" ++ show y ++ ")"
+
+tokenToPair :: L.Token a -> (Ann, a)
+tokenToPair t = (ann t, L.payload t)
+
+newtype Input = Input { inputStream :: [L.Token L.Lexeme] }
+  deriving (Eq, Ord, Show)
+
+instance P.Stream Input where
+  type Token Input = L.Token L.Lexeme
+  type Tokens Input = Input
+
+  tokenToChunk pxy = P.tokensToChunk pxy . pure
+
+  tokensToChunk _ = Input
+
+  chunkToTokens _ = inputStream
+
+  chunkLength pxy = length . P.chunkToTokens pxy
+
+  chunkEmpty pxy = null . P.chunkToTokens pxy
+
+  positionAt1 _ sp t = setPos sp (L.start t)
+
+  positionAtN pxy sp =
+    fromMaybe sp . fmap (setPos sp . L.start) . listToMaybe . P.chunkToTokens pxy
+
+  advance1 _ _ cp = setPos cp . L.end
+
+  advanceN _ _ cp = setPos cp . L.end . last . inputStream
+
+  take1_ (P.chunkToTokens proxy -> []) = Nothing
+  take1_ (P.chunkToTokens proxy -> t:ts) = Just (t, P.tokensToChunk proxy ts)
+  take1_ _ = error "Unpossible"
+
+  takeN_ n (P.chunkToTokens proxy -> []) | n > 0 = Nothing
+  takeN_ n ts =
+    Just
+      . join bimap (P.tokensToChunk proxy)
+      . splitAt n $ P.chunkToTokens proxy ts
+
+  takeWhile_ p = join bimap (P.tokensToChunk proxy) . span p . inputStream
+
+setPos :: P.SourcePos -> L.Pos -> P.SourcePos
+setPos sp lp =
+  P.SourcePos (P.sourceName sp) (P.mkPos $ L.line lp) (P.mkPos $ L.column lp)
+
+class Annotated a where
+  ann :: a -> Ann
+
+instance Annotated Ann where
+  ann = id
+
+instance Annotated (L.Token a) where
+  ann (L.Token _ s e) = Ann s e
+
+instance Annotated a => Annotated (ABT.Term f v a) where
+  ann = ann . ABT.annotation
+
+instance Annotated a => Annotated (PatternP a) where
+  ann = ann . Pattern.loc
+
+instance (Annotated a, Annotated b) => Annotated (MatchCase a b) where
+  ann (MatchCase p _ b) = ann p <> ann b
+
+mkAnn :: (Annotated a, Annotated b) => a -> b -> Ann
+mkAnn x y = ann x <> ann y
+
+showLineCol :: Annotated a => a -> String
+showLineCol a =
+  let L.Pos line col = start $ ann a
+  in "Line " ++ show line ++ ", column " ++ show col
+
+tok :: (Ann -> a -> b) -> L.Token a -> b
+tok f (L.Token a start end) = f (Ann start end) a
+
+peekAny :: Var v => P v (L.Token L.Lexeme)
+peekAny = P.lookAhead P.anyChar
+
+proxy :: Proxy Input
+proxy = Proxy
+
+root :: Var v => P v a -> P v a
+root p = p <* P.eof
+
+run' :: P v a -> String -> String -> PEnv -> Either (Err v) a
+run' p s name = runParserT p name (Input $ L.lexer name s) -- todo: L.reorder
+
+run :: P v a -> String -> PEnv -> Either (Err v) a
+run p s = run' p s ""
 
 penv0 :: PEnv
 penv0 = Map.empty
-type Parser s a = Parsec.ParsecT Text (Env s) ((->) PEnv) a
 
-data Env s = Env s L.LayoutEnv
+-- Virtual pattern match on a lexeme.
+queryToken :: Var v => (L.Lexeme -> Maybe a) -> P v (L.Token a)
+queryToken f = P.token go Nothing
+  where go t@((f . L.payload) -> Just s) = Right $ fmap (const s) t
+        go x = Left (pure (P.Tokens (x:|[])), Set.empty)
 
-instance L.HasLayoutEnv (Env s) where
-  getLayoutEnv (Env _ l) = l
-  setLayoutEnv l (Env s _) = Env s l
-
-pTrace :: [Char] -> Text.Parsec.Prim.ParsecT Text.Text (Env s) ((->) PEnv) ()
-pTrace s = pt <|> return ()
-    where pt = attempt $
-               do
-                 x <- attempt $ many anyChar
-                 void $ trace (s++": " ++show x) $ attempt $ char 'z'
-                 fail x
-
-tracingEnabled :: Bool
-tracingEnabled = False
-
-traced :: [Char]
-       -> Text.Parsec.Prim.ParsecT Text.Text (Env s) ((->) PEnv) b
-       -> Text.Parsec.Prim.ParsecT Text.Text (Env s) ((->) PEnv) b
-traced s p = if not tracingEnabled then p else Parsec.parserTraced s p
-
-root :: Parser s a -> Parser s a
-root p = optional (L.space) *> (p <* (optional semicolon <* eof))
-
-semicolon :: Parser s ()
-semicolon = void $ L.spaced L.semi
-
-semicolon2 :: Parser s ()
-semicolon2 = semicolon *> semicolon
-
-eof :: Parser s ()
-eof = Parsec.eof
-
-attempt :: Parser s a -> Parser s a
-attempt = Parsec.try
-
-lookAhead :: Parser s a -> Parser s a
-lookAhead = Parsec.lookAhead
-
-run' :: Parser s a
-     -> String
-     -> s
-     -> String
-     -> PEnv
-     -> Either String a
-run' p s s0 name =
-  err <$> Parsec.runParserT p (Env s0 L.defaultLayoutEnv) name (Text.pack s)
-  where err (Left e) = Left (show e)
-        err (Right a) = Right a
-
-run :: Parser s a
-    -> String
-    -> s
-    -> PEnv
-    -> Either String a
-run p s s0 = run' p s s0 ""
-
-unsafeRun :: Parser s a -> String -> s -> PEnv -> a
-unsafeRun p s s0 = err <$> run p s s0
-  where err (Right a) = a
-        err (Left e) = error e
-
-string :: String -> Parser s String
-string s = attempt (Parsec.string s) <|> fail ("expected '" ++ s ++ "'")
-
-takeLine :: String -> String
-takeLine = Prelude.takeWhile (/= '\n')
-
-char :: Char -> Parser s Char
-char = Parsec.char
-
-one :: String -> (Char -> Bool) -> Parser s Char
-one msg f = attempt $ Parsec.anyChar >>= \ch -> case f ch of
-  True -> pure ch
-  False -> fail $ "invalid " ++ msg ++ " '" ++ [ch] ++ "'"
-
-base64string' :: String -> Parser s String
-base64string' alphabet = concat <$> many base64group
+-- Consume a block opening and return the string that opens the block.
+openBlock :: Var v => P v (L.Token String)
+openBlock = queryToken getOpen
   where
-    base64group :: Parser s String
-    base64group = do
-      chars <- some $ one "expected base64 character" (`elem` alphabet)
-      padding <- sequenceA (replicate (padCount $ length chars) (char '='))
-      return $ chars ++ padding
-    padCount :: Int -> Int
-    padCount len = case len `mod` 4 of 0 -> 0; n -> 4 - n
+    getOpen (L.Open s) = Just s
+    getOpen _ = Nothing
 
-base64urlstring :: Parser s String
-base64urlstring = base64string' $ ['A' .. 'Z'] ++ ['a' .. 'z'] ++ ['0' .. '9'] ++ "-_"
+openBlockWith :: Var v => String -> P v (L.Token ())
+openBlockWith s = fmap (const ()) <$> P.satisfy ((L.Open s ==) . L.payload)
 
-notReservedChar :: Char -> Bool
-notReservedChar = (`notElem` "\".,`[]{}:()@")
+withinBlock :: Var v => P v a -> P v a
+withinBlock p = openBlock *> p <* closeBlock
 
-identifier :: String -> [String -> Bool] -> Parser s String
-identifier msg = identifier' msg [not . isSpace, notReservedChar]
+-- Match a particular lexeme exactly, and consume it.
+matchToken :: Var v => L.Lexeme -> P v (L.Token L.Lexeme)
+matchToken x = P.satisfy ((==) x . L.payload)
 
-label :: String -> Parser s a -> Parser s a
-label msg p = Parsec.label p msg
+-- Consume a virtual semicolon
+semi :: Var v => P v (L.Token ())
+semi = fmap (const ()) <$> matchToken L.Semi
 
-identifier' :: String -> [Char -> Bool] -> [String -> Bool] -> Parser s String
-identifier' msg charTests stringTests = attempt $ do
-  i <- takeWhile1 msg (\c -> all ($ c) charTests)
-  guard (all ($ i) stringTests)
-  pure i
+-- Consume the end of a block
+closeBlock :: Var v => P v (L.Token ())
+closeBlock = fmap (const ()) <$> matchToken L.Close
 
--- a wordyId isn't all digits, isn't all symbols, and isn't a symbolyId
-wordyId :: [String] -> Parser s String
-wordyId keywords = label "wordyId" . token $ do
-  op <- (False <$ symbolyId keywords) <|> pure True
-  guard op
-  f <$> sepBy1 dot id -- todo: this screws up ∀ a. without a space following 'a'
+-- Parse an alphanumeric identifier
+wordyId :: Var v => P v (L.Token String)
+wordyId = queryToken getWordy
+  where getWordy (L.WordyId s) = Just s
+        getWordy _ = Nothing
+
+-- Parse a symboly ID like >>= or &&
+symbolyId :: Var v => P v (L.Token String)
+symbolyId = queryToken getSymboly
+  where getSymboly (L.SymbolyId s) = Just s
+        getSymboly _ = Nothing
+
+backticks :: Var v => P v (L.Token String)
+backticks = queryToken getBackticks
+  where getBackticks (L.Backticks s) = Just s
+        getBackticks _ = Nothing
+
+-- Parse a reserved word
+reserved :: Var v => String -> P v (L.Token String)
+reserved w = P.label w $ queryToken getReserved
+  where getReserved (L.Reserved w') | w == w' = Just w
+        getReserved _ = Nothing
+
+numeric :: Var v => P v (L.Token String)
+numeric = queryToken getNumeric
+  where getNumeric (L.Numeric s) = Just s
+        getNumeric _ = Nothing
+
+sepBy :: Var v => P v a -> P v b -> P v [b]
+sepBy sep pb = P.sepBy pb sep
+
+sepBy1 :: Var v => P v a -> P v b -> P v [b]
+sepBy1 sep pb = P.sepBy1 pb sep
+
+prefixVar :: Var v => P v (L.Token v)
+prefixVar = fmap (Var.named . Text.pack) <$> P.label "symbol" prefixOp
   where
-    dot = attempt (char '.')
-    id = identifier "alphanumeric identifier"
-      [any (not . Char.isDigit), any Char.isAlphaNum, (`notElem` keywords)]
-    f segs = intercalate "." segs
+    prefixOp = wordyId <|> P.label "prefix-operator" (P.try (reserved "(" *> symbolyId) <* reserved ")")
 
--- a symbolyId is all symbols
-symbolyId :: [String] -> Parser s String
-symbolyId keywords = label "operator" . token $ do
-  op <- identifier' "operator identifier"
-    [notReservedChar,
-     (/= '_'),
-     not . Char.isSpace,
-     \c -> Char.isSymbol c || Char.isPunctuation c]
-    [(`notElem` keywords)]
-  qual <- optional (char '_' *> wordyId keywords)
-  pure $ maybe op (\qual -> qual ++ "." ++ op) qual
+infixVar :: Var v => P v (L.Token v)
+infixVar =
+  fmap (Var.named . Text.pack) <$> (symbolyId <|> backticks)
 
-token :: Parser s a -> Parser s a
-token p = attempt (L.spaced p)
+hashLiteral :: Var v => P v (L.Token Hash)
+hashLiteral = queryToken getHash
+  where getHash (L.Hash s) = Just s
+        getHash _ = Nothing
 
-token_ :: Parser s a -> Parser s ()
-token_ = void . token
+string :: Var v => P v (L.Token Text)
+string = queryToken getString
+  where getString (L.Textual s) = Just (Text.pack s)
+        getString _ = Nothing
 
-parenthesized :: Parser s a -> Parser s a
-parenthesized p = lp *> body <* rp
+tupleOrParenthesized :: Var v => P v a -> (Ann -> a) -> (a -> a -> a) -> P v a
+tupleOrParenthesized p unit pair = do
+    open <- reserved "("
+    es <- sepBy (reserved ",") p
+    close <- optional semi *> reserved ")"
+    pure $ go es open close
   where
-    lp = char '(' <* L.withoutLayout "space" (optional L.space)
-    body = L.withoutLayout "parentheses" p
-    rp = token (char ')')
+    go [t] _ _ = t
+    go as s e = foldr pair (unit (ann s <> ann e)) as
 
-takeWhile :: String -> (Char -> Bool) -> Parser s String
-takeWhile msg f = label msg (Parsec.many (one msg f))
-
-takeWhile1 :: String -> (Char -> Bool) -> Parser s String
-takeWhile1 msg f = label msg (Parsec.many1 (one msg f))
-
-whitespace :: Parser s ()
-whitespace = void (optional L.space)
-
-whitespace1 :: Parser s ()
-whitespace1 = void L.space
-
-chainl1 :: Parser s a -> Parser s (a -> a -> a) -> Parser s a
-chainl1 = Parsec.chainl1
-
-chainl :: Parser s a -> Parser s (a -> a -> a) -> a -> Parser s a
-chainl = Parsec.chainl
-
-sepBy :: Parser s a -> Parser s b -> Parser s [b]
-sepBy sep pb = Parsec.sepBy pb sep
-
-sepBy1 :: Parser s a -> Parser s b -> Parser s [b]
-sepBy1 sep pb = Parsec.sepBy1 pb sep
-
-get :: Parser s s
-get = (\(Env s _) -> s) <$> Parsec.getState
-
-set :: s -> Parser s ()
-set s = do
-  Env _ l <- Parsec.getState
-  Parsec.putState (Env s l)
+chainl1 :: Var v => P v a -> P v (a -> a -> a) -> P v a
+chainl1 p op = foldl (flip ($)) <$> p <*> P.many (flip <$> op <*> p)
