+++ parser-typechecker/src/Unison/TermParser.hs
@@ -1,35 +1,34 @@
+{-# LANGUAGE ViewPatterns #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE RankNTypes #-}
-{-# Language OverloadedStrings #-}
-{-# Language ScopedTypeVariables #-}
-{-# Language BangPatterns #-}
-{-# Language TupleSections #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE BangPatterns #-}
+{-# LANGUAGE TupleSections #-}
 
 module Unison.TermParser where
 
 import           Control.Applicative
-import           Control.Monad
+import           Control.Monad (guard, join, when, void)
 import           Control.Monad.Reader (ask)
-import           Data.Char (isDigit, isUpper)
-import           Data.Foldable (asum,toList)
-import           Data.Functor
+import           Data.Char (isUpper)
+import           Data.Foldable (asum)
 import           Data.Int (Int64)
+import           Data.List (elem)
 import qualified Data.Map as Map
 import           Data.Maybe (isJust)
-import qualified Data.Text as Text
 import           Data.Word (Word64)
-import           Debug.Trace
-import           Prelude hiding (and, or, takeWhile)
-import qualified Text.Parsec.Layout as L
+import           Prelude hiding (and, or)
+import qualified Text.Megaparsec as P
 import qualified Unison.ABT as ABT
+import qualified Unison.Lexer as L
 import           Unison.Parser
-import           Unison.Pattern (Pattern)
-import qualified Unison.Pattern as Pattern
+import           Unison.PatternP (Pattern)
+import qualified Unison.PatternP as Pattern
 import qualified Unison.Reference as R
-import           Unison.Term (Term)
+import           Unison.Term (AnnotatedTerm)
 import qualified Unison.Term as Term
-import           Unison.Type (Type)
-import qualified Unison.Type as Type
+import           Unison.Type (AnnotatedType)
 import qualified Unison.TypeParser as TypeParser
 import           Unison.Var (Var)
 import qualified Unison.Var as Var
@@ -44,9 +43,7 @@ operator characters (like empty? or fold-left).
 Sections / partial application of infix operators is not implemented.
 -}
 
-type S = TypeParser.S
-
-type TermP v = Parser (S v) (Term v)
+type TermP v = P v (AnnotatedTerm v Ann)
 
 term :: Var v => TermP v
 term = term2
@@ -57,10 +54,10 @@ term2 = lam term2 <|> term3
 term3 :: Var v => TermP v
 term3 = do
   t <- letBlock <|> handle <|> ifthen <|> and <|> or <|> match <|> infixApp
-  ot <- optional (token (char ':') *> TypeParser.valueType)
+  ot <- optional (reserved ":" *> TypeParser.computationType)
   pure $ case ot of
     Nothing -> t
-    Just y -> Term.ann() t y
+    Just y -> Term.ann (mkAnn t y) t y
 
 -- We disallow type annotations and lambdas,
 -- just function application and operators
@@ -70,294 +67,254 @@ blockTerm = letBlock <|> handle <|> ifthen <|> and <|> or <|> match <|>
 
 match :: Var v => TermP v
 match = do
-  token_ $ string "case"
+  start <- reserved "case"
   scrutinee <- term
-  token_ $ string "of"
-  cases <- L.vblockNextToken (sepBy L.vsemi matchCase)
-  pure $ Term.match() scrutinee cases
+  _ <- openBlockWith "of"
+  -- TODO: Produce a nice error message for empty match list
+  cases <- sepBy1 semi matchCase
+  _ <- closeBlock
+  pure $ Term.match (ann start <> ann (last cases)) scrutinee cases
 
-matchCase :: Var v => Parser (S v) (Term.MatchCase () (Term v))
+matchCase :: Var v => P v (Term.MatchCase Ann (AnnotatedTerm v Ann))
 matchCase = do
   (p, boundVars) <- parsePattern
-  guard <- traced "guard" $ optional $ token (string "|") *> infixApp
-  traced "->" $ token_ $ string "->"
-  t <- block
-  pure . Term.MatchCase p guard $ ABT.absChain boundVars t
+  guard <- optional $ reserved "|" *> infixApp
+  t <- block "->"
+  pure . Term.MatchCase p guard $ ABT.absChain' boundVars t
 
-parsePattern :: Var v => Parser (S v) (Pattern, [v])
-parsePattern = traced "pattern" $ constructor <|> leaf
+parsePattern :: forall v. Var v => P v (Pattern Ann, [(Ann, v)])
+parsePattern = constructor <|> leaf
   where
   leaf = literal <|> varOrAs <|> unbound <|>
          parenthesizedOrTuplePattern <|> effect
-  literal = traced "pattern.literal" $ (,[]) <$> asum [true, false, number]
-  true = Pattern.Boolean True <$ token (string "true")
-  false = Pattern.Boolean False <$ token (string "false")
-  number = traced "pattern.number" $ number' Pattern.Int64 Pattern.UInt64 Pattern.Float
+  literal = (,[]) <$> asum [true, false, number]
+  true = (\t -> Pattern.Boolean (ann t) True) <$> reserved "true"
+  false = (\t -> Pattern.Boolean (ann t) False) <$> reserved "false"
+  number = number' (tok Pattern.Int64) (tok Pattern.UInt64) (tok Pattern.Float)
+  parenthesizedOrTuplePattern :: P v (Pattern Ann, [(Ann, v)])
   parenthesizedOrTuplePattern = tupleOrParenthesized parsePattern unit pair
-  unit = (Pattern.Constructor (R.Builtin "()") 0 [], [])
+  unit ann = (Pattern.Constructor ann (R.Builtin "()") 0 [], [])
   pair (p1, v1) (p2, v2) =
-    (Pattern.Constructor (R.Builtin "Pair") 0 [p1, p2], v1 ++ v2)
-  varOrAs = traced "varOrAs" $ do
+    (Pattern.Constructor (ann p1 <> ann p2) (R.Builtin "Pair") 0 [p1, p2],
+     v1 ++ v2)
+  varOrAs :: P v (Pattern Ann, [(Ann, v)])
+  varOrAs = do
     v <- prefixVar
-    o <- optional (token $ string "@")
+    o <- optional (reserved "@")
     if isJust o then
-      (\(p, vs) -> (Pattern.As p, v : vs)) <$> leaf
-      else pure (Pattern.Var, [v])
-  unbound = traced "unbound" $ (Pattern.Unbound, []) <$ token (char '_')
-  ctorName = traced "ctorName" . token $ do
-    s <- wordyId keywords
-    guard . isUpper . head $ s
+      (\(p, vs) -> (Pattern.As (ann v) p, tokenToPair v : vs)) <$> leaf
+      else pure (Pattern.Var (ann v), [tokenToPair v])
+  unbound :: P v (Pattern Ann, [(Ann, v)])
+  unbound = (\tok -> (Pattern.Unbound (ann tok), [])) <$> reserved "_"
+  ctorName = P.try $ do
+    s <- wordyId
+    guard . isUpper . head . L.payload $ s
     pure s
 
-  effectBind0 = traced "effectBind0" $ do
+  effectBind0 = do
     name <- ctorName
     leaves <- many leaf
-    token_ (string "->")
+    _ <- reserved "->"
     pure (name, leaves)
 
   effectBind = do
-    (name, leaves) <- attempt effectBind0
+    (name, leaves) <- P.try effectBind0
     (cont, vsp) <- parsePattern
     env <- ask
-    (ref,cid) <- case Map.lookup name env of
+    (ref,cid) <- case Map.lookup (L.payload name) env of
       Just (ref, cid) -> pure (ref, cid)
-      Nothing -> fail $ "unknown data constructor " ++ name
+      Nothing -> customFailure $ UnknownEffectConstructor name
     pure $ case unzip leaves of
       (patterns, vs) ->
-         (Pattern.EffectBind ref cid patterns cont, join vs ++ vsp)
+         (Pattern.EffectBind (ann name <> ann cont) ref cid patterns cont,
+          join vs ++ vsp)
 
   effectPure = go <$> parsePattern where
-    go (p, vs) = (Pattern.EffectPure p, vs)
+    go (p, vs) = (Pattern.EffectPure (ann p) p, vs)
 
   effect = do
-    token_ (string "{")
-    (effectBind <|> effectPure) <* token_ (string "}")
+    start <- openBlockWith "{"
+    (inner, vs) <- effectBind <|> effectPure
+    _ <- closeBlock
+    end <- reserved "}"
+    pure $ (Pattern.setLoc inner (ann start <> ann end), vs)
 
-  constructor = traced "constructor" $ do
-    name <- ctorName
+  constructor = do
+    t <- ctorName
+    let name = L.payload t
     env <- ask
     case Map.lookup name env of
-      Just (ref, cid) -> go <$> traced "pattern.manyleaf" (many leaf)
+      Just (ref, cid) -> go <$> many leaf
         where
           go pairs = case unzip pairs of
-            (patterns, vs) -> (Pattern.Constructor ref cid patterns, join vs)
-      Nothing ->
-        trace ("unknown data constructor " ++ name) $
-          (traced ("failing " ++ name) . fail $ "unknown data constructor " ++ name)
+            (patterns, vs) ->
+              let loc = foldl (<>) (ann t) $ map ann patterns
+              in (Pattern.Constructor loc ref cid patterns, join vs)
+      Nothing -> customFailure $ UnknownDataConstructor t
 
-letBlock :: Var v => TermP v
-letBlock = traced "letBlock" $ token (string "let") *> block
-
-infixApp :: Var v => TermP v
-infixApp = chainl1 term4 (f <$> infixVar)
+lam :: Var v => TermP v -> TermP v
+lam p = P.label "lambda" $ mkLam <$> P.try (some prefixVar <* reserved "->") <*> p
   where
-    f :: Ord v => v -> Term v -> Term v -> Term v
-    f op lhs rhs = Term.apps (Term.var() op) [((),lhs),((),rhs)]
+    mkLam vs b = Term.lam' (ann (head vs) <> ann b) (map L.payload vs) b
 
-term4 :: Var v => TermP v
-term4 = traced "apply-chain" $ f <$> some termLeaf
+letBlock, handle, ifthen, and, or, infixApp :: Var v => TermP v
+letBlock = P.label "let" $ block "let"
+
+handle = P.label "handle" $ do
+  t <- reserved "handle"
+  handler <- term
+  b <- block "in"
+  pure $ Term.handle (ann t <> ann b) handler b
+
+ifthen = P.label "if" $ do
+  c <- block "if"
+  t <- block "then"
+  f <- block "else"
+  pure $ Term.iff (ann c <> ann f) c t f
+
+hashLit :: Var v => TermP v
+hashLit = tok Term.derived <$> hashLiteral
+
+prefixTerm :: Var v => TermP v
+prefixTerm = tok Term.var <$> prefixVar
+
+text :: Var v => TermP v
+text = tok Term.text <$> string
+
+boolean :: Var v => TermP v
+boolean = ((\t -> Term.boolean (ann t) True) <$> reserved "true") <|>
+          ((\t -> Term.boolean (ann t) False) <$> reserved "false")
+
+blank :: Var v => TermP v
+blank = (\t -> Term.blank (ann t)) <$> reserved "_"
+
+vector :: Var v => TermP v -> TermP v
+vector p = f <$> reserved "[" <*> elements <*> reserved "]"
   where
-    f (func:args) = Term.apps func (((),) <$> args)
-    f [] = error "'some' shouldn't produce an empty list"
+    elements = sepBy (reserved ",") p
+    f open elems close = Term.vector (ann open <> ann close) elems
 
 termLeaf :: forall v. Var v => TermP v
-termLeaf = traced "leaf" $
-  asum [hashLit, prefixTerm, text, number, boolean, tupleOrParenthesizedTerm, blank, vector term]
+termLeaf =
+  asum [hashLit, prefixTerm, text, number, boolean,
+        tupleOrParenthesizedTerm, blank, vector term]
 
-ifthen :: Var v => TermP v
-ifthen = traced "ifthen" $ do
-  token_ $ string "if"
-  cond <- block' $ L.virtual_rbrace <|> (lookAhead . token_ $ string "then")
-  token_ $ string "then"
-  iftrue <- block' $ L.virtual_rbrace <|> (lookAhead . token_ $ string "else")
-  token_ $ string "else"
-  iffalse <- block
-  pure $ Term.iff() cond iftrue iffalse
+and = P.label "and" $ f <$> reserved "and" <*> termLeaf <*> termLeaf
+  where f kw x y = Term.and (ann kw <> ann y) x y
 
-and :: Var v => TermP v
-and = Term.and() <$> (token (string "and") *> termLeaf) <*> termLeaf
+or = P.label "or" $ f <$> reserved "or" <*> termLeaf <*> termLeaf
+  where f kw x y = Term.or (ann kw <> ann y) x y
 
-or :: Var v => TermP v
-or = Term.or() <$> (token (string "or") *> termLeaf) <*> termLeaf
+var :: Var v => L.Token v -> AnnotatedTerm v Ann
+var t = Term.var (ann t) (L.payload t)
 
--- Generic parser for tuples and parenthesized patterns/terms
-tupleOrParenthesized :: Parser s a -> a -> (a -> a -> a) -> Parser s a
-tupleOrParenthesized rec unit pair =
-  parenthesized $ go <$> sepBy (token $ string ",") rec where
-    go [t] = t -- was just a parenthesized term
-    go terms = foldr pair unit terms -- it's a tuple literal
-
--- Specialized to terms
-tupleOrParenthesizedTerm :: Var v => TermP v
-tupleOrParenthesizedTerm = tupleOrParenthesized term unit pair
+term4 :: Var v => TermP v
+term4 = f <$> some termLeaf
   where
-    pair t1 t2 =
-      Term.constructor() (R.Builtin "Pair") 0 `Term.app_` t1 `Term.app_` t2
-    unit = Term.constructor() (R.Builtin "()") 0
+    f (func:args) = Term.apps func ((\a -> (ann func <> ann a, a)) <$> args)
+    f [] = error "'some' shouldn't produce an empty list"
 
-text' :: Parser s Text.Text
-text' =
-  token $ fmap Text.pack ps
-  where ps = char '"' *> Unison.Parser.takeWhile "text literal" (/= '"') <* char '"'
-
-text :: Ord v => Parser s (Term v)
-text = Term.text() <$> text'
-
-number :: Ord v => Parser s (Term v)
-number = number' (Term.int64()) (Term.uint64()) (Term.float())
-
-number' :: (Int64 -> a) -> (Word64 -> a) -> (Double -> a) -> Parser s a
-number' i u f = token $ do
-  let digits = takeWhile1 "number" isDigit
-  sign <- optional (char '+' <|> char '-')
-  ds <- digits
-  fraction <- optional ((:) <$> char '.' <*> digits)
-  pure $ case fraction of
-    Nothing -> case sign of
-      Nothing -> u (read ds)
-      Just '+' -> i (read ds)
-      Just '-' -> i (read ('-':ds))
-      _ -> error "impossible!"
-    Just fraction ->
-      let signl = toList sign
-      in f (read (signl ++ ds ++ fraction))
-
-boolean :: Ord v => Parser s (Term v)
-boolean =
-  (Term.boolean() True <$ token (string "true")) <|>
-  (Term.boolean() False <$ token (string "false"))
-
-hashLit :: Ord v => Parser s (Term v)
-hashLit = token (f =<< (mark *> hash))
+infixApp = chainl1 term4 (f <$> fmap var infixVar)
   where
-    f h = case Term.derived' (Text.pack h) of
-      Nothing -> fail "invalid base58 string"
-      Just a -> pure a
-    mark = char '#'
-    hash = base64urlstring
+    f op lhs rhs =
+      Term.apps op [(ann lhs, lhs), (ann rhs, rhs)]
 
-blank :: Ord v => TermP v
-blank = token (char '_') $> Term.blank()
+typedecl :: Var v => P v (L.Token v, AnnotatedType v Ann)
+typedecl =
+  (,) <$> P.try (prefixVar <* reserved ":")
+      <*> TypeParser.valueType
+      <* semi
 
-vector :: Ord v => TermP v -> TermP v
-vector p = Term.vector() <$> (lbracket *> elements <* rbracket)
-  where
-    lbracket = token (char '[')
-    elements = sepBy comma (L.withoutLayout "vector element" p)
-    comma = token (char ',')
-    rbracket = token (char ']')
-
-eq :: Parser (S v) ()
-eq = void $ char '=' <* L.space
-
-binding :: Var v => Parser (S v) (v, Term v)
-binding = traced "binding" . label "binding" $ do
+binding :: forall v. Var v => P v ((Ann, v), AnnotatedTerm v Ann)
+binding = P.label "binding" $ do
   typ <- optional typedecl
-  let lhs = attempt ((\arg1 op arg2 -> (op,[arg1,arg2]))
-                    <$> prefixVar <*> infixVar <*> prefixVar)
-                <|> ((,) <$> prefixVar <*> many prefixVar)
+  let infixLhs = do
+        (arg1, op) <- P.try ((,) <$> prefixVar <*> infixVar)
+        arg2 <- prefixVar
+        pure (ann arg1, L.payload op, [arg1, arg2])
+  let prefixLhs = do
+        v  <- prefixVar
+        vs <- many prefixVar
+        pure (ann v, L.payload v, vs)
+  let
+    lhs :: P v (Ann, v, [L.Token v])
+    lhs = infixLhs <|> prefixLhs
   case typ of
     Nothing -> do
       -- we haven't seen a type annotation, so lookahead to '=' before commit
-      (name, args) <- attempt (lhs <* eq)
-      body <- block
-      pure $ mkBinding name args body
+      (loc, name, args) <- P.try (lhs <* P.lookAhead (openBlockWith "="))
+      body <- block "="
+      pure $ mkBinding loc name args body
     Just (nameT, typ) -> do
-      (name, args) <- lhs
-      when (name /= nameT) $
-        fail ("The type signature for ‘" ++ show (Var.name nameT) ++ "’ lacks an accompanying binding")
-      body <- eq *> block
-      pure $ fmap (\e -> Term.ann () e typ) (mkBinding name args body)
+      (_, name, args) <- lhs
+      when (name /= L.payload nameT) $
+        customFailure $ SignatureNeedsAccompanyingBody nameT
+      body <- block "="
+      pure $ fmap (\e -> Term.ann (ann nameT <> ann e) e typ)
+                  (mkBinding (ann nameT) name args body)
   where
-  mkBinding f [] body = (f, body)
-  mkBinding f args body = (f, Term.lam' () args body)
+  mkBinding loc f [] body = ((loc, f), body)
+  mkBinding loc f args body =
+    ((loc, f), Term.lam' (loc <> ann body) (L.payload <$> args) body)
 
-typedecl :: Var v => Parser (S v) (v, Type v)
-typedecl = (,) <$> attempt (prefixVar <* token (char ':'))
-               <*> L.vblockIncrement TypeParser.valueType
 
-infixVar :: Var v => Parser s v
-infixVar = (Var.named . Text.pack) <$> (backticked <|> symbolyId keywords)
+customFailure :: P.MonadParsec e s m => e -> m a
+customFailure = P.customFailure
+
+block :: forall v. Var v => String -> TermP v
+block s = block' s (openBlockWith s) closeBlock
+
+topBlock :: forall v. Var v => TermP v
+topBlock = block' "top-level block" (void <$> peekAny) closeBlock
+
+block' :: forall v. Var v => String -> P v (L.Token ()) -> P v (L.Token ()) -> TermP v
+block' s openBlock closeBlock = do
+    open <- openBlock
+    statements <- sepBy semi statement
+    _ <- closeBlock
+    go open statements
   where
-    backticked = attempt (char '`') *> wordyId keywords <* token (char '`')
+    statement = (Right <$> binding) <|> (Left <$> blockTerm)
+    toBinding (Right ((a, v), e)) = ((a, v), e)
+    toBinding (Left e) = ((ann e, Var.named "_"), e)
+    go :: L.Token () -> [Either (AnnotatedTerm v Ann) ((Ann, v), AnnotatedTerm v Ann)] -> P v (AnnotatedTerm v Ann)
+    go open bs =
+      let startAnnotation = (fst . fst . toBinding $ head bs)
+      in case reverse bs of
+        Right ((a, _v), annotatedTerm) : _ ->
+          customFailure $ BlockMustEndWithExpression
+                            (startAnnotation <> ann annotatedTerm)
+                            (a <> ann annotatedTerm)
+        Left e : es ->
+          pure $ Term.letRec (startAnnotation <> ann e)
+                             (toBinding <$> reverse es)
+                             e
+        [] -> customFailure $ EmptyBlock (const s <$> open)
 
-prefixVar :: Var v => Parser s v
-prefixVar = traced "prefixVar" $ (Var.named . Text.pack) <$> label "symbol" (token prefixOp)
+
+number :: Var v => TermP v
+number = number' (tok Term.int64) (tok Term.uint64) (tok Term.float)
+
+number'
+  :: Var v
+  => (L.Token Int64 -> a)
+  -> (L.Token Word64 -> a)
+  -> (L.Token Double -> a)
+  -> P v a
+number' i u f = fmap go numeric
   where
-    prefixOp = wordyId keywords
-           <|> (char '(' *> symbolyId keywords <* token (char ')')) -- no whitespace w/in parens
+    go num@(L.payload -> p)
+      | elem '.' p = f (read <$> num)
+      | take 1 p == "+" || take 1 p == "-" = i (read <$> num)
+      | otherwise = u (read <$> num)
 
-prefixTerm :: Var v => TermP v
-prefixTerm = Term.var() <$> prefixVar
-
-keywords :: [String]
-keywords =
-  [ "->"
-  , ":"
-  , "="
-  , "let"
-  , "alias"
-  , "and", "or"
-  , "case", "of"
-  , "handle", "in"
-  , "if", "then", "else"
-  , "namespace"
-  , "type", "effect", "where"
-  , "true", "false"
-  ]
-
-block' :: Var v => Parser (S v) () -> TermP v
-block' vendbrace =
-  traced "block" $
-    go =<< L.vblock' L.virtual_lbrace_nextToken vendbrace (sepBy L.vsemi statement)
+tupleOrParenthesizedTerm :: Var v => TermP v
+tupleOrParenthesizedTerm = P.label "tuple" $ tupleOrParenthesized term unit pair
   where
-    statement =
-      traced "statement" $ (Right <$> binding) <|> (Left <$> blockTerm)
-    toBinding (Right (v, e)) = (v,e)
-    toBinding (Left e) = (Var.named "_", e)
-    go bs = case reverse bs of
-      (Right _e : _) -> fail "block must end with an expression"
-      -- TODO: Inform the user that we're going to rewrite the block,
-      -- possibly changing the meaning of the program (which is ambiguous anyway),
-      -- or fail with a helpful error message if there's a forward reference with
-      -- effects.
-      (Left e : bs) -> pure $ Term.letRec_ (toBinding <$> reverse bs) e
-      [] -> fail "empty block"
-
-block :: Var v => TermP v
-block = block' (traced "L.virtual_rbrace" L.virtual_rbrace)
-
-handle :: Var v => TermP v
-handle = traced "handle" $ do
-  token_ $ string "handle"
-  handler <- term
-  token_ $ string "in"
-  b <- block
-  pure $ Term.handle() handler b
-
-lam :: Var v => TermP v -> TermP v
-lam p = traced "lambda" $ attempt (Term.lam' () <$> vars <* arrow) <*> body
-  where
-    vars = some prefixVar
-    arrow = token (string "->")
-    body = p
-
-alias :: Var v => Parser (S v) ()
-alias = do
-  _ <- token (string "alias")
-  (fn:params) <- some (Var.named . Text.pack <$> wordyId keywords)
-  _ <- eq
-  body <- L.vblockIncrement TypeParser.valueType
-  TypeParser.Aliases s <- get
-  let s' = (fn, apply)
-      apply args | length args <= length params = ABT.substs (params `zip` args) body
-      apply args = apply (take n args) `Type.apps` (addUnit <$> drop n args)
-      n = length params
-      addUnit a = ((), a)
-  set (TypeParser.Aliases (s':s))
-
--- bindings :: Var v => Parser (S v) [(v, Term v)]
--- bindings = do s0 <- get; L.laidout (many alias *> binding) <* set s0 where
-
--- moduleBindings :: Var v => Parser (S v) [(v, Term v)]
--- moduleBindings = root bindings
+    pair t1 t2 =
+      Term.app (ann t1 <> ann t2)
+        (Term.app (ann t1)
+                  (Term.constructor (ann t1 <> ann t2) (R.Builtin "Pair") 0)
+                  t1)
+        t2
+    unit ann = Term.constructor ann (R.Builtin "()") 0
