--- parser-typechecker/src/Unison/TermParser2.hs
@@ -1,320 +1,0 @@
-{-# LANGUAGE ViewPatterns #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE RankNTypes #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE TupleSections #-}
-
-module Unison.TermParser2 where
-
-import           Control.Applicative
-import           Control.Monad (guard, join, when, void)
-import           Control.Monad.Reader (ask)
-import           Data.Char (isUpper)
-import           Data.Foldable (asum)
-import           Data.Int (Int64)
-import           Data.List (elem)
-import qualified Data.Map as Map
-import           Data.Maybe (isJust)
-import           Data.Word (Word64)
-import           Prelude hiding (and, or)
-import qualified Text.Megaparsec as P
-import qualified Unison.ABT as ABT
-import qualified Unison.Lexer as L
-import           Unison.Parser2
-import           Unison.PatternP (Pattern)
-import qualified Unison.PatternP as Pattern
-import qualified Unison.Reference as R
-import           Unison.Term (AnnotatedTerm)
-import qualified Unison.Term as Term
-import           Unison.Type (AnnotatedType)
-import qualified Unison.TypeParser2 as TypeParser
-import           Unison.Var (Var)
-import qualified Unison.Var as Var
-
-{-
-Precedence of language constructs is identical to Haskell, except that all
-operators (like +, <*>, or any sequence of non-alphanumeric characters) are
-left-associative and equal precedence, and operators must have surrounding
-whitespace (a + b, not a+b) to distinguish from identifiers that may contain
-operator characters (like empty? or fold-left).
-
-Sections / partial application of infix operators is not implemented.
--}
-
-type TermP v = P v (AnnotatedTerm v Ann)
-
-term :: Var v => TermP v
-term = term2
-
-term2 :: Var v => TermP v
-term2 = lam term2 <|> term3
-
-term3 :: Var v => TermP v
-term3 = do
-  t <- letBlock <|> handle <|> ifthen <|> and <|> or <|> match <|> infixApp
-  ot <- optional (reserved ":" *> TypeParser.computationType)
-  pure $ case ot of
-    Nothing -> t
-    Just y -> Term.ann (mkAnn t y) t y
-
--- We disallow type annotations and lambdas,
--- just function application and operators
-blockTerm :: Var v => TermP v
-blockTerm = letBlock <|> handle <|> ifthen <|> and <|> or <|> match <|>
-            lam term <|> infixApp
-
-match :: Var v => TermP v
-match = do
-  start <- reserved "case"
-  scrutinee <- term
-  _ <- openBlockWith "of"
-  -- TODO: Produce a nice error message for empty match list
-  cases <- sepBy1 semi matchCase
-  _ <- closeBlock
-  pure $ Term.match (ann start <> ann (last cases)) scrutinee cases
-
-matchCase :: Var v => P v (Term.MatchCase Ann (AnnotatedTerm v Ann))
-matchCase = do
-  (p, boundVars) <- parsePattern
-  guard <- optional $ reserved "|" *> infixApp
-  t <- block "->"
-  pure . Term.MatchCase p guard $ ABT.absChain' boundVars t
-
-parsePattern :: forall v. Var v => P v (Pattern Ann, [(Ann, v)])
-parsePattern = constructor <|> leaf
-  where
-  leaf = literal <|> varOrAs <|> unbound <|>
-         parenthesizedOrTuplePattern <|> effect
-  literal = (,[]) <$> asum [true, false, number]
-  true = (\t -> Pattern.Boolean (ann t) True) <$> reserved "true"
-  false = (\t -> Pattern.Boolean (ann t) False) <$> reserved "false"
-  number = number' (tok Pattern.Int64) (tok Pattern.UInt64) (tok Pattern.Float)
-  parenthesizedOrTuplePattern :: P v (Pattern Ann, [(Ann, v)])
-  parenthesizedOrTuplePattern = tupleOrParenthesized parsePattern unit pair
-  unit ann = (Pattern.Constructor ann (R.Builtin "()") 0 [], [])
-  pair (p1, v1) (p2, v2) =
-    (Pattern.Constructor (ann p1 <> ann p2) (R.Builtin "Pair") 0 [p1, p2],
-     v1 ++ v2)
-  varOrAs :: P v (Pattern Ann, [(Ann, v)])
-  varOrAs = do
-    v <- prefixVar
-    o <- optional (reserved "@")
-    if isJust o then
-      (\(p, vs) -> (Pattern.As (ann v) p, tokenToPair v : vs)) <$> leaf
-      else pure (Pattern.Var (ann v), [tokenToPair v])
-  unbound :: P v (Pattern Ann, [(Ann, v)])
-  unbound = (\tok -> (Pattern.Unbound (ann tok), [])) <$> reserved "_"
-  ctorName = P.try $ do
-    s <- wordyId
-    guard . isUpper . head . L.payload $ s
-    pure s
-
-  effectBind0 = do
-    name <- ctorName
-    leaves <- many leaf
-    _ <- reserved "->"
-    pure (name, leaves)
-
-  effectBind = do
-    (name, leaves) <- P.try effectBind0
-    (cont, vsp) <- parsePattern
-    env <- ask
-    (ref,cid) <- case Map.lookup (L.payload name) env of
-      Just (ref, cid) -> pure (ref, cid)
-      Nothing -> customFailure $ UnknownEffectConstructor name
-    pure $ case unzip leaves of
-      (patterns, vs) ->
-         (Pattern.EffectBind (ann name <> ann cont) ref cid patterns cont,
-          join vs ++ vsp)
-
-  effectPure = go <$> parsePattern where
-    go (p, vs) = (Pattern.EffectPure (ann p) p, vs)
-
-  effect = do
-    start <- openBlockWith "{"
-    (inner, vs) <- effectBind <|> effectPure
-    _ <- closeBlock
-    end <- reserved "}"
-    pure $ (Pattern.setLoc inner (ann start <> ann end), vs)
-
-  constructor = do
-    t <- ctorName
-    let name = L.payload t
-    env <- ask
-    case Map.lookup name env of
-      Just (ref, cid) -> go <$> many leaf
-        where
-          go pairs = case unzip pairs of
-            (patterns, vs) ->
-              let loc = foldl (<>) (ann t) $ map ann patterns
-              in (Pattern.Constructor loc ref cid patterns, join vs)
-      Nothing -> customFailure $ UnknownDataConstructor t
-
-lam :: Var v => TermP v -> TermP v
-lam p = P.label "lambda" $ mkLam <$> P.try (some prefixVar <* reserved "->") <*> p
-  where
-    mkLam vs b = Term.lam' (ann (head vs) <> ann b) (map L.payload vs) b
-
-letBlock, handle, ifthen, and, or, infixApp :: Var v => TermP v
-letBlock = P.label "let" $ block "let"
-
-handle = P.label "handle" $ do
-  t <- reserved "handle"
-  handler <- term
-  b <- block "in"
-  pure $ Term.handle (ann t <> ann b) handler b
-
-ifthen = P.label "if" $ do
-  c <- block "if"
-  t <- block "then"
-  f <- block "else"
-  pure $ Term.iff (ann c <> ann f) c t f
-
-hashLit :: Var v => TermP v
-hashLit = tok Term.derived <$> hashLiteral
-
-prefixTerm :: Var v => TermP v
-prefixTerm = tok Term.var <$> prefixVar
-
-text :: Var v => TermP v
-text = tok Term.text <$> string
-
-boolean :: Var v => TermP v
-boolean = ((\t -> Term.boolean (ann t) True) <$> reserved "true") <|>
-          ((\t -> Term.boolean (ann t) False) <$> reserved "false")
-
-blank :: Var v => TermP v
-blank = (\t -> Term.blank (ann t)) <$> reserved "_"
-
-vector :: Var v => TermP v -> TermP v
-vector p = f <$> reserved "[" <*> elements <*> reserved "]"
-  where
-    elements = sepBy (reserved ",") p
-    f open elems close = Term.vector (ann open <> ann close) elems
-
-termLeaf :: forall v. Var v => TermP v
-termLeaf =
-  asum [hashLit, prefixTerm, text, number, boolean,
-        tupleOrParenthesizedTerm, blank, vector term]
-
-and = P.label "and" $ f <$> reserved "and" <*> termLeaf <*> termLeaf
-  where f kw x y = Term.and (ann kw <> ann y) x y
-
-or = P.label "or" $ f <$> reserved "or" <*> termLeaf <*> termLeaf
-  where f kw x y = Term.or (ann kw <> ann y) x y
-
-var :: Var v => L.Token v -> AnnotatedTerm v Ann
-var t = Term.var (ann t) (L.payload t)
-
-term4 :: Var v => TermP v
-term4 = f <$> some termLeaf
-  where
-    f (func:args) = Term.apps func ((\a -> (ann func <> ann a, a)) <$> args)
-    f [] = error "'some' shouldn't produce an empty list"
-
-infixApp = chainl1 term4 (f <$> fmap var infixVar)
-  where
-    f op lhs rhs =
-      Term.apps op [(ann lhs, lhs), (ann rhs, rhs)]
-
-typedecl :: Var v => P v (L.Token v, AnnotatedType v Ann)
-typedecl =
-  (,) <$> P.try (prefixVar <* reserved ":")
-      <*> TypeParser.valueType
-      <* semi
-
-binding :: forall v. Var v => P v ((Ann, v), AnnotatedTerm v Ann)
-binding = P.label "binding" $ do
-  typ <- optional typedecl
-  let infixLhs = do
-        (arg1, op) <- P.try ((,) <$> prefixVar <*> infixVar)
-        arg2 <- prefixVar
-        pure (ann arg1, L.payload op, [arg1, arg2])
-  let prefixLhs = do
-        v  <- prefixVar
-        vs <- many prefixVar
-        pure (ann v, L.payload v, vs)
-  let
-    lhs :: P v (Ann, v, [L.Token v])
-    lhs = infixLhs <|> prefixLhs
-  case typ of
-    Nothing -> do
-      -- we haven't seen a type annotation, so lookahead to '=' before commit
-      (loc, name, args) <- P.try (lhs <* P.lookAhead (openBlockWith "="))
-      body <- block "="
-      pure $ mkBinding loc name args body
-    Just (nameT, typ) -> do
-      (_, name, args) <- lhs
-      when (name /= L.payload nameT) $
-        customFailure $ SignatureNeedsAccompanyingBody nameT
-      body <- block "="
-      pure $ fmap (\e -> Term.ann (ann nameT <> ann e) e typ)
-                  (mkBinding (ann nameT) name args body)
-  where
-  mkBinding loc f [] body = ((loc, f), body)
-  mkBinding loc f args body =
-    ((loc, f), Term.lam' (loc <> ann body) (L.payload <$> args) body)
-
-
-customFailure :: P.MonadParsec e s m => e -> m a
-customFailure = P.customFailure
-
-block :: forall v. Var v => String -> TermP v
-block s = block' s (openBlockWith s) closeBlock
-
-topBlock :: forall v. Var v => TermP v
-topBlock = block' "top-level block" (void <$> peekAny) closeBlock
-
-block' :: forall v. Var v => String -> P v (L.Token ()) -> P v (L.Token ()) -> TermP v
-block' s openBlock closeBlock = do
-    open <- openBlock
-    statements <- sepBy semi statement
-    _ <- closeBlock
-    go open statements
-  where
-    statement = (Right <$> binding) <|> (Left <$> blockTerm)
-    toBinding (Right ((a, v), e)) = ((a, v), e)
-    toBinding (Left e) = ((ann e, Var.named "_"), e)
-    go :: L.Token () -> [Either (AnnotatedTerm v Ann) ((Ann, v), AnnotatedTerm v Ann)] -> P v (AnnotatedTerm v Ann)
-    go open bs =
-      let startAnnotation = (fst . fst . toBinding $ head bs)
-      in case reverse bs of
-        Right ((a, _v), annotatedTerm) : _ ->
-          customFailure $ BlockMustEndWithExpression
-                            (startAnnotation <> ann annotatedTerm)
-                            (a <> ann annotatedTerm)
-        Left e : es ->
-          pure $ Term.letRec (startAnnotation <> ann e)
-                             (toBinding <$> reverse es)
-                             e
-        [] -> customFailure $ EmptyBlock (const s <$> open)
-
-
-number :: Var v => TermP v
-number = number' (tok Term.int64) (tok Term.uint64) (tok Term.float)
-
-number'
-  :: Var v
-  => (L.Token Int64 -> a)
-  -> (L.Token Word64 -> a)
-  -> (L.Token Double -> a)
-  -> P v a
-number' i u f = fmap go numeric
-  where
-    go num@(L.payload -> p)
-      | elem '.' p = f (read <$> num)
-      | take 1 p == "+" || take 1 p == "-" = i (read <$> num)
-      | otherwise = u (read <$> num)
-
-tupleOrParenthesizedTerm :: Var v => TermP v
-tupleOrParenthesizedTerm = P.label "tuple" $ tupleOrParenthesized term unit pair
-  where
-    pair t1 t2 =
-      Term.app (ann t1 <> ann t2)
-        (Term.app (ann t1)
-                  (Term.constructor (ann t1 <> ann t2) (R.Builtin "Pair") 0)
-                  t1)
-        t2
-    unit ann = Term.constructor ann (R.Builtin "()") 0
