+++ parser-typechecker/src/Unison/TypeParser.hs
@@ -1,39 +1,23 @@
-{-# Language OverloadedStrings #-}
-{-# Language BangPatterns #-}
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Unison.TypeParser where
 
-import Control.Monad
-import Control.Applicative
-import Data.Char (isUpper, isLower)
-import Data.List
-import Unison.Parser
-import Unison.Type (Type)
-import Unison.Var (Var)
+import           Control.Applicative
+import           Control.Monad
+import           Data.Char (isUpper, isLower)
+import           Data.List
 import qualified Data.Text as Text
+import qualified Text.Megaparsec as P
+import qualified Unison.Lexer as L
+import           Unison.Parser
+import           Unison.Type (AnnotatedType)
 import qualified Unison.Type as Type
+import           Unison.Var (Var)
 
---import Debug.Trace
---
---pTrace s = pt <|> return ()
---    where pt = attempt $
---               do
---                 x <- attempt $ many anyChar
---                 trace (s++": " ++x) $ attempt $ char 'z'
---                 fail x
---
----- traced s p = p
---traced s p = do
---  pTrace s
---  a <- p <|> trace (s ++ " backtracked") (fail s)
---  let !x = trace (s ++ " succeeded") ()
---  pure a
-
-newtype S v = Aliases [(v, [Type v] -> Type v)]
-s0 :: S v
-s0 = Aliases []
-
-type TypeP v = Parser (S v) (Type v)
+-- A parsed type is annotated with its starting and ending position in the
+-- source text.
+type TypeP v = P v (AnnotatedType v Ann)
 
 -- Value types cannot have effects, unless those effects appear to
 -- the right of a function arrow:
@@ -48,10 +32,10 @@ computationType = effect <|> valueType
 
 valueTypeLeaf :: Var v => TypeP v
 valueTypeLeaf =
-  tupleOrParenthesized valueType <|> typeVar
+  tupleOrParenthesizedType valueType <|> typeVar
 
 typeVar :: Var v => TypeP v
-typeVar = fmap (Type.v' . Text.pack) (token $ wordyId keywords)
+typeVar = posMap (\pos -> Type.av' pos . Text.pack) wordyId
 
 type1 :: Var v => TypeP v
 type1 = arrow type2
@@ -62,68 +46,63 @@ type2 = app valueTypeLeaf
 -- ex : {State Text, IO} (Sequence Int64)
 effect :: Var v => TypeP v
 effect = do
-  token_ $ string "{"
-  es <- sepBy (token (string ",")) valueType
-  token_ $ string "}"
+  open <- reserved "{"
+  es <- sepBy (reserved ",") valueType
+  _ <- reserved "}"
   t <- valueTypeLeaf
-  pure (Type.effect() es t)
+  pure (Type.effect (Ann (L.start open) (end $ ann t)) es t)
 
-tupleOrParenthesized :: Ord v => TypeP v -> TypeP v
-tupleOrParenthesized rec =
-  parenthesized $ go <$> sepBy (token $ string ",") rec where
-    go [t] = t
-    go types = foldr pair unit types
-    pair t1 t2 = Type.app() (Type.app() (Type.builtin() "Pair") t1) t2
-    unit = Type.builtin() "()"
+tupleOrParenthesizedType :: Var v => TypeP v -> TypeP v
+tupleOrParenthesizedType rec = tupleOrParenthesized rec unit pair
+  where
+    pair t1 t2 =
+      let a = ann t1 <> ann t2
+      in Type.app a (Type.app (ann t1) (Type.builtin a "Pair") t1) t2
+    unit a = Type.builtin a "()"
 
 -- "TypeA TypeB TypeC"
 app :: Ord v => TypeP v -> TypeP v
-app rec = get >>= \(Aliases aliases) -> do
+app rec = do
   (hd:tl) <- some rec
-  pure $ case hd of
-    Type.Var' v -> case lookup v aliases of
-      Nothing -> foldl' (Type.app()) hd tl
-      Just apply -> apply tl
-    _ -> foldl' (Type.app()) hd tl
+  pure $ foldl' (\a b -> Type.app (ann a <> ann b) a b) hd tl
 
 --  valueType ::= ... | Arrow valueType computationType
 arrow :: Var v => TypeP v -> TypeP v
-arrow rec = do
-  t <- foldr1 (Type.arrow()) <$> sepBy1 (token (string "->")) (effect <|> rec)
-  case t of
-    Type.Arrow' (Type.Effect' _ _) _ -> fail "effect to the left of an ->"
-    _ -> pure t
+arrow rec =
+  let p = sepBy1 (reserved "->") (effect <|> rec)
+  in foldr1 (\a b -> Type.arrow (ann a <> ann b) a b) <$> p
 
 -- "forall a b . List a -> List b -> Maybe Text"
 forall :: Var v => TypeP v -> TypeP v
 forall rec = do
-    (void . token $ string "forall") <|> void (token (char '∀'))
-    vars <- some $ token varName
-    _ <- token (char '.')
+    kw <- reserved "forall" <|> reserved "∀"
+    vars <- fmap (fmap L.payload) . some $ varName
+    _ <- matchToken $ L.SymbolyId "."
     t <- rec
-    pure $ Type.forall'() (fmap Text.pack vars) t
+    pure $ Type.forall' (ann kw <> ann t) (fmap Text.pack vars) t
 
-varName :: Parser s String
+varName :: Var v => P v (L.Token String)
 varName = do
-  name <- wordyId keywords
-  guard (isLower . head $ name)
+  name <- wordyId
+  guard (isLower . head $ L.payload name)
   pure name
 
-typeName :: Parser s String
+typeName :: Var v => P v (L.Token String)
 typeName = do
-  name <- wordyId keywords
-  guard (isUpper . head $ name)
+  name <- wordyId
+  guard (isUpper . head $ L.payload name)
   pure name
 
-keywords :: [String]
-keywords = ["forall", "∀"]
-
--- qualifiedTypeName :: Parser String
+-- qualifiedTypeName :: P v (L.Token String
 -- qualifiedTypeName = f <$> typeName <*> optional more
 --   where
 --     f :: String -> (Maybe String) -> String
 --     f first more = maybe first (first++) more
 --     more = (:) <$> char '.' <*> qualifiedTypeName
 
+posMap :: (Ann -> a -> b) -> P v (L.Token a) -> P v b
+posMap f = fmap $ \case L.Token a start end -> f (Ann start end) a
+
 literal :: Var v => TypeP v
-literal = label "literal" . token $ (Type.v' . Text.pack) <$> typeName
+literal =
+  P.label "literal" . posMap (\pos -> Type.av' pos . Text.pack) $ typeName
