+++ parser-typechecker/tests/Unison/Test/Common.hs
@@ -3,20 +3,24 @@ module Unison.Test.Common where
 import qualified Data.Map as Map
 import qualified Unison.Builtin as B
 import qualified Unison.FileParsers as FP
-import           Unison.Symbol (Symbol)
-import           Unison.Term (Term)
-import           Unison.Type (Type)
-import qualified Unison.Typechecker as Typechecker
+import           Unison.Parser (Ann(..))
+import           Unison.Result (Result,Note)
 import qualified Unison.Result as Result
-import Unison.Result (Result,Note)
+import           Unison.Symbol (Symbol)
+import           Unison.Term (AnnotatedTerm)
+import           Unison.Type (AnnotatedType)
+import qualified Unison.Typechecker as Typechecker
+
+type Term v = AnnotatedTerm v Ann
+type Type v = AnnotatedType v Ann
 
 tm :: String -> Term Symbol
 tm = B.tm
 
-file :: String -> Result (Note Symbol ()) (Term Symbol, Type Symbol)
+file :: String -> Result (Note Symbol Ann) (Term Symbol, Type Symbol)
 file = FP.parseAndSynthesizeAsFile ""
 
-fileTerm :: String -> Result (Note Symbol ()) (Term Symbol)
+fileTerm :: String -> Result (Note Symbol Ann) (Term Symbol)
 fileTerm = fmap fst . FP.parseAndSynthesizeAsFile "<test>"
 
 fileTermType :: String -> Maybe (Term Symbol, Type Symbol)
@@ -28,8 +32,8 @@ t = B.t
 typechecks :: String -> Bool
 typechecks = Result.isSuccess . file
 
-env :: Monad m => Typechecker.Env m Symbol ()
-env = Typechecker.Env () [] typeOf dd ed where
+env :: Monad m => Typechecker.Env m Symbol Ann
+env = Typechecker.Env Intrinsic [] typeOf dd ed where
   typeOf r = maybe (error $ "no type for: " ++ show r) pure $ Map.lookup r B.builtins
   dd r = error $ "no data declaration for: " ++ show r
   ed r = error $ "no effect declaration for: " ++ show r
